% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{check_normality}
\alias{check_normality}
\title{Check normality}
\usage{
check_normality(data, vars, group = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Numeric columns as character names.}

\item{group}{Optional grouping column name.}

\item{alpha}{Significance level.}
}
\description{
Check normality
}
\keyword{internal}
