% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{check_variance_homogeneity}
\alias{check_variance_homogeneity}
\title{Check homogeneity of variance}
\usage{
check_variance_homogeneity(data, outcome, group, alpha = 0.05)
}
\description{
Check homogeneity of variance
}
\keyword{internal}
