% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{check_variance_two_groups}
\alias{check_variance_two_groups}
\title{Check variance for two groups}
\usage{
check_variance_two_groups(data, outcome, group, alpha = 0.05)
}
\description{
Check variance for two groups
}
\keyword{internal}
