% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives_categorical}
\alias{descriptives_categorical}
\title{Categorical descriptive statistics}
\usage{
descriptives_categorical(data, vars)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Categorical columns.}
}
\description{
Categorical descriptive statistics
}
\keyword{internal}
