% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives_numeric}
\alias{descriptives_numeric}
\title{Numeric descriptive statistics}
\usage{
descriptives_numeric(data, vars, group = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Numeric columns.}

\item{group}{Optional grouping column.}
}
\description{
Numeric descriptive statistics
}
\keyword{internal}
