% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.testflow}
\alias{plot.testflow}
\title{Plot a testflow object}
\usage{
\method{plot}{testflow}(x, title = NULL, subtitle = NULL, caption = NULL, ...)
}
\arguments{
\item{x}{A testflow object.}

\item{title}{Optional plot title override. Defaults to the stored title.}

\item{subtitle}{Optional plot subtitle override. Defaults to the stored
subtitle.}

\item{caption}{Optional plot caption override. Defaults to the stored
caption.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object stored in the \code{testflow} object, with optional
title, subtitle, and caption overrides applied. If the workflow was created
with \code{plot = FALSE}, returns \code{NULL}.
}
\description{
Plot a testflow object
}
