% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{print.summary.sample_size}
\alias{print.summary.sample_size}
\title{Print a summary for a sample size object}
\usage{
\method{print}{summary.sample_size}(x, ...)
}
\arguments{
\item{x}{A \code{summary.sample_size} object.}

\item{...}{Unused.}
}
\value{
The object invisibly.
}
\description{
Print a summary for a sample size object
}
