% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Return a ready-to-use testflow report}
\usage{
report(x, ...)
}
\arguments{
\item{x}{A testflow object.}

\item{...}{Unused.}
}
\description{
Return a ready-to-use testflow report
}
