% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_test}
\alias{report_test}
\title{Return a ready-to-use testflow report}
\usage{
report_test(x)
}
\arguments{
\item{x}{A testflow object.}
}
\value{
A length-one character vector containing the same report text as
\code{\link[=report]{report()}} for a \code{testflow} object.
}
\description{
Return a ready-to-use testflow report
}
