% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_correlation.R
\name{test_correlation_matrix}
\alias{test_correlation_matrix}
\title{Test a correlation matrix}
\usage{
test_correlation_matrix(
  data,
  vars,
  method = c("spearman", "pearson", "kendall"),
  alpha = 0.05,
  plot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Numeric columns.}

\item{method}{Correlation method.}

\item{alpha}{Significance level.}

\item{plot}{Logical; include a ggplot object.}

\item{na.rm}{Logical; remove missing values.}
}
\description{
Test a correlation matrix
}
