% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multinomial.R
\name{test_multinomial}
\alias{test_multinomial}
\title{Test multinomial goodness of fit}
\usage{
test_multinomial(
  data,
  outcome,
  p = NULL,
  alpha = 0.05,
  plot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{outcome}{Categorical outcome column.}

\item{p}{Expected probabilities, or \code{NULL} for equal probabilities.}

\item{alpha}{Significance level.}

\item{plot}{Logical; include a ggplot object.}

\item{na.rm}{Logical; remove missing values.}
}
\description{
Test multinomial goodness of fit
}
