% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\title{Convert a testflow object to a one-row tibble}
\usage{
as_tibble(x, ...)
}
\arguments{
\item{x}{A testflow object.}

\item{...}{Unused.}
}
\description{
Convert a testflow object to a one-row tibble
}
