% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.testflow}
\alias{print.testflow}
\title{Print a testflow object}
\usage{
\method{print}{testflow}(x, ...)
}
\arguments{
\item{x}{A testflow object.}

\item{...}{Unused.}
}
\description{
Print a testflow object.
}
\details{
Console colors are enabled by default in interactive sessions. Use
\code{options(testflow.cli_colors = FALSE)} to disable colors, or
\code{options(testflow.cli_colors = TRUE)} to force colors in non-interactive
output.
}
