% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.testflow}
\alias{summary.testflow}
\title{Summarize a testflow object}
\usage{
\method{summary}{testflow}(object, ...)
}
\arguments{
\item{object}{A testflow object.}

\item{...}{Unused.}
}
\description{
Summarize a testflow object
}
