% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_paired.R
\name{test_paired}
\alias{test_paired}
\title{Compare paired before and after numeric measurements}
\usage{
test_paired(
  data,
  before,
  after,
  alternative = c("two.sided", "less", "greater"),
  alpha = 0.05,
  plot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{before}{Before column.}

\item{after}{After column.}

\item{alternative}{Alternative hypothesis.}

\item{alpha}{Significance level.}

\item{plot}{Logical; include a ggplot object.}

\item{na.rm}{Logical; remove missing values.}
}
\description{
Compare paired before and after numeric measurements
}
