% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_paired_categorical.R
\name{test_paired_categorical}
\alias{test_paired_categorical}
\title{Test paired categorical measurements}
\usage{
test_paired_categorical(
  data,
  before,
  after,
  alpha = 0.05,
  plot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{before}{Before categorical column.}

\item{after}{After categorical column.}

\item{alpha}{Significance level.}

\item{plot}{Logical; include a ggplot object.}

\item{na.rm}{Logical; remove missing values.}
}
\description{
Test paired categorical measurements
}
