% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_proportion.R
\name{test_proportion}
\alias{test_proportion}
\title{Test a one-sample proportion}
\usage{
test_proportion(
  data,
  outcome,
  success,
  p = 0.5,
  alpha = 0.05,
  plot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{outcome}{Categorical outcome column.}

\item{success}{Value counted as success.}

\item{p}{Reference probability.}

\item{alpha}{Significance level.}

\item{plot}{Logical; include a ggplot object.}

\item{na.rm}{Logical; remove missing values.}
}
\description{
Test a one-sample proportion
}
